requirements.txt
aiogram~=3.12.0
python-dotenv~=1.0.0
utils~=1.0.2